Fix Infrastructure namespace Rename SetResetToken -> SetResetDisposable
